+++ src/lib.rs
@@ -42,11 +42,9 @@ doc_comment::doctest!("../README.md");
 mod activation_factory;
 mod agile_object;
 mod array;
-mod bstring;
 mod com_interface;
 mod com_ptr;
 mod error;
-mod error_info;
 #[doc(hidden)]
 pub mod factory;
 mod guid;
@@ -54,7 +52,6 @@ mod hstring;
 mod object;
 mod param;
 mod ref_count;
-mod restricted_error_info;
 pub mod runtime;
 mod runtime_name;
 mod runtime_type;
